Move Poco to its own class file
